+++ src/main/java/org/chzz/market/domain/auction/dto/event/AuctionImageUpdateEvent.java
@@ -1,4 +1,4 @@
-package org.chzz.market.domain.auction.dto;
+package org.chzz.market.domain.auction.dto.event;
 
 import java.util.Map;
 import org.chzz.market.domain.auction.entity.Auction;
